Create 9c3Find the Rotation Count in Rotated Sorted array
Given an array arr[] of size N having distinct numbers sorted in increasing order and the array has been right rotated (i.e, the last element will be cyclically shifted to the starting position of the array) k number of times, the task is to find the value of k.

Examples:  

Input: arr[] = {15, 18, 2, 3, 6, 12}
Output: 2
Explanation: Initial array must be {2, 3, 6, 12, 15, 18}. 
We get the given array after rotating the initial array twice.

Input: arr[] = {7, 9, 11, 12, 5}
Output: 4
Approach 2: (Efficient Using Binary Search) 

A better approach would be to perform binary searching, in place of linear search to find the position of the smallest element in the array.

Follow the steps mentioned below to implement the idea: 

The minimum element is the only element whose previous is greater than it. If there is no previous element, then there is no rotation (the first element is minimum). 
Check this condition for the middle element by comparing it with the (mid-1)th and (mid+1)th elements.
If the minimum element is not at the middle (neither mid nor mid + 1), then 
If the middle element is smaller than the last element, then the minimum element lies in the left half
Else minimum element lies in the right half.
Illustration:

Let the array be arr[]={15, 18, 2, 3, 6, 12}
low = 0 , high = 5.
            =>  mid = 2
            =>  arr[mid]=2 , arr[mid-1] > arr[mid] , hence condition is matched
            =>  The required index = mid = 2

So the element is  found at index 2.

Below is the implementation of the above approach.
Time Complexity: O(log N)