Create 11Find the Minimum element in a Sorted and Rotated Array
Given a sorted array arr[] (may be distinct or may contain duplicates) of size N that is rotated at some unknown point, the task is to find the minimum element in it. 

Examples: 

Input: arr[] = {5, 6, 1, 2, 3, 4}
Output: 1
Explanation: 1 is the minimum element present in the array.

Input: arr[] = {1, 2, 3, 4}
Output: 1

Find the minimum element in a sorted and rotated array using Binary Search: 
This approach is based on the following idea:

As the array is sorted and rotated, there are two segments that are themselves sorted but their meeting point is the only position where the smallest element is and that is not sorted. 
Follow the steps below to solve the given problem: 

If we take a closer look at the above examples, we can easily figure out the following pattern:

The minimum element is the only element whose previous is greater than it. If there is no previous element, then there is no rotation (the first element is minimum). 
We check this condition for the middle element by comparing it with (mid-1)th and (mid+1)th elements.
If the minimum element is not at the middle (neither mid nor mid + 1), then: 
If the middle element is smaller than the last element, then the minimum element lies in the left half
Else minimum element lies in the right half.
Follow the below illustration for a better understanding

Illustration:

Let the array be arr[]={15, 18, 2, 3, 6, 12}
low = 0 , high = 5.
            =>  mid = 2
            =>  arr[mid]=2 , arr[mid-1] > arr[mid] , hence condition is matched
            =>  The required index = mid = 2

So the element is  found at index 2 and arr[2] = 2
Time Complexity: O(logN)