Create 12Find a rotation with maximum hamming distance
Given an array of N elements, create a new array which is a rotation of the given array, and hamming distance between both the arrays is maximum. The task is to print the maximized hamming distance between given and new arrays.

Hamming distance between two arrays or strings of equal length is the number of positions at which the corresponding characters (elements) are different.

Examples: 

Input: N = 3, arr = {1, 4, 1}
Output: 2
Explanation:  Possible rotations of given array = 4 1 1 and 1 1 4. In each case the hamming distance is 2. Therefore the maximum hamming distance will be 2.
Naive Approach: 
The idea is to create another array that is double the size of the original array, such that the elements of this new array (copy array) are just the elements of the original array repeated twice in the same sequence. For example, if the original array is 1 4 1, then the copy array is 1 4 1 1 4 1. Now, iterate through the copy array and find hamming distance with every shift (or rotation). So we check 4 1 1, 1 1 4, 1 4 1, and choose the output for which the hamming distance is maximum.

Illustration:

Given array arr[]={2, 4, 6, 8}.
New array brr[]={2, ,4, 6, 8, 2, 4, 6, 8} , count=0

In first iteration : {2, 4, 6, 8} & {2 ,4, 6, 8, 2, 4, 6, 8} , count=1
In second iteration : {2, 4, 6, 8} & {2 ,4, 6, 8, 2, 4, 6, 8} , count=2
In third iteration : {2, 4, 6, 8} & {2 ,4, 6, 8, 2, 4, 6, 8} , count=3
In fourth iteration : {2, 4, 6, 8} & {2 ,4, 6, 8, 2, 4, 6, 8} , count=4
count = size of original array, hence output is 4
Implementation:

We will create another array twice the size of the original array and insert the elements in it one by one, twice. Now we will perform rotation on the array and for each rotation, we will check if the value of the original array matches the newly created array.

If they don’t match we will increase our counter value
After the value is incremented we will perform a condition to check the max value, to keep a check on the maximum value that can be obtained.
Else if the values don’t match, we don’t have to perform any operation
Inside the checking loop, we will also have to keep a check if the value of counter = size of the original array, as it is the maximum value that we can obtain. If the condition matches, we may return the value. This condition optimizes our code.
Time Complexity: O(n2)