Create 12c2Find a rotation with maximum hamming distance
Examples: 

Input: N = 3, arr = {1, 4, 1}
Output: 2
Explanation:  Possible rotations of given array = 4 1 1 and 1 1 4. In each case the hamming distance is 2. Therefore the maximum hamming distance will be 2.


Approach 2 (in Constant Space): 
The idea is to compare elements of the original array sequence with its rotated versions. The rotated versions of the array are achieved using shifted index method where you compare elements at the original index with elements on the shifted index, without requiring any extra space.

Illustration:

Given array arr[]={2, 4, 6, 8}. count=0. We will use modulo operation to compare the two different indexes of the same array.

In first iteration : {2, 4, 6, 8} & {2 ,4, 6, 8} , count=1
In second iteration : {2, 4, 6, 8} & {2 ,4, 6, 8} , count=2
In third iteration : {2, 4, 6, 8} & {2 ,4, 6, 8} , count=3
In fourth iteration : {2, 4, 6, 8} & {2 ,4, 6, 8} , count=4
count = size of original array, hence output is 4
Follow the steps below for the above idea:

We will perform rotation on the array and for each rotation, we will check if the value of the array at any index i ,  matches with the value of index j. Where j determines the rotation and i as the actual index for which the comparison has to be made. 

j = 1,2,3 … , n-1 & i= 0, 1, 2 , … ,n-1,  n is the size of our array.
If arr[i] not equal to arr[(i+j)%n] we will increase our counter value
Else if the values don’t match, we don’t have to perform any operation
Inside the checking loop, we will also have to keep a check if the value of counter = size of the original array, as it is the maximum value that we can obtain. If the condition matches, we may return the value. This condition optimizes our code.
Below is the implementation of the above approach: