Create 16c2Find element at given index after a number of rotations
Given an array of N integers, M ranges of indices as queries, and a specific index K, the task is to right rotate the array circularly between query ranges, and return the element at Kth index in the final array. 

Examples: 

Input: arr[] : {1, 2, 3, 4, 5}, M = 2, queries[] = { {0, 2}, {0, 3} }, K = 1
Output: 3
Explanation: After rotation according to 1st query {0, 2} = arr[] = {3, 1, 2, 4, 5}
After rotation according to 2nd query {0, 3} = arr[] = {4, 3, 1, 2, 5}
Now after M rotations, the element at Kth index = arr[1] = 3

Input: arr[] : {1, 2, 3, 4, 5}, M = 2, queries[] = { {0, 3}, {3, 4} }, K = 1
Output: 1
Explanation: After rotation according to 1st query {0, 3} = arr[] = {4, 1, 2, 3, 5}
After rotation according to 2nd query {3, 4} = arr[] = {4, 1, 2, 5, 3}
Now after M rotations, the element at Kth index = arr[1] = 1
Efficient Approach:

The idea is to calculate the relative position of the index each time the rotation is performed. We can do this simply by checking for the range in which the rotation is to be performed and updating the value of the index according to that. But we need to perform the rotation in a reverse manner. 

The reason for doing so can be justified as follows:

It may be possible that we are rotating some overlapped positions repeatedly.
In those cases, we are going deeper into the previous rotations starting from the latest rotation. 
So to get the proper position we are iterating from the reverse direction.
Please note that we are not rotating the elements in the reverse order, just processing the index from the reverse. Because if we actually rotate in reverse order, we might get a completely different answer as in the case of rotations the order matters.

Follow along with the below illustration for a better understanding.
Illustration:

Let the given array be arr[] = {10, 20, 30, 40, 50 }
query: {{0, 2} {0, 3}}, K = 1. 

We will consider the rotations from the backside.

query 2 {0, 3}:
        => index = 1 and index ≥ 0 and index ≤ 3
        => index ≠ left therefore, index = 0

query 1 {0, 2}:
        => index = 0 and index >= 0 and index <= 2
        => index = left therefore, index = 2

Therefore we return arr[index] =arr[2] = 30

Follow the below steps to implement the idea:

Iterate the queries from reverse direction:
If the range contains the value K, then update its position before rotation.
Continue the above two steps till all the queries are traversed.
Return the value at the index K (K updated through all the queries)
Below is the implementation of the above approach: